computeSimpleReimbursementPlan to behaviour
